+++ src/zeus2_toolbox/tools.py
@@ -394,9 +394,198 @@ def build_header(header_dict):
                  list(header_dict.items())[0][1][-1][-1])
     else:
         header_str = "%s_%04d-%s_%04d" % \
-                        (list(header_dict.items())[0][0],
-                         list(header_dict.items())[0][1][0][0],
-                         list(header_dict.items())[-1][0],
-                         list(header_dict.items())[-1][1][-1][-1])
+                     (list(header_dict.items())[0][0],
+                      list(header_dict.items())[0][1][0][0],
+                      list(header_dict.items())[-1][0],
+                      list(header_dict.items())[-1][1][-1][-1])
 
     return header_str
+
+
+def spec_to_wl(spec, spat, grat_idx, order=5, rd=200, rg=0.711111111111111,
+               px_shift=0, py_shift=0, alpha_min_index=73.9062453,
+               c0=0.9800910119, c1=-0.0017009051035, c2=-0.87654448327,
+               c3=36.248043521, c4=459.42373214, c5=-80.04474108,
+               c6=-0.0017003774252, c7=-1.5498032937, c8=102.04705483):
+    """
+    get the wavelength corresponding to the given grating index, spatial
+    position, spectral index and the fitted parameters for ZEUS2; the default
+    value for the fitted parameters are taken from
+    zeus2_grating_calibration_apex2019 file; wavelength is calculated as
+
+        lambda = 5/order * (a (sin alpha_s)^2 + b * sin alpha_s + c)
+
+    where alpha_s = alpha_min_index - grat_idx / (rd * rg) and
+    theta = px + c6 * py**2 + c7 * py + c8; px_shift and py_shift can shift
+    lab calibration to fit with sky calibration
+
+    :param spec: int or int array, spectral index used to compute wavelength; if
+        input spec is an array, it must be compatible with spat so that
+        spat+spec does not raise an error
+    :type spec: Union[int, numpy.ndarray]
+    :param spat: int or int array, spatial position considered; if input spat is
+        an array, it must be compatible with spec so that spat+spec does not
+        raise an error, and the output array shape will be the same as the
+        shape of spat+spec array
+    :type spat: Union[int, numpy.ndarray]
+    :param int grat_idx: int, grating index
+    :param int order: int, grating order to use
+    :param rd: int or float, stepper motor, degrees per step
+    :type rd: Union[int, float]
+    :param float rg: float, grating degrees of movement per rotation of drive
+        shaft
+    :param px_shift: int or float, shift for spatial position to reconcile sky
+        wavelength with fit parameters, will be added to spat to get px in theta
+    :type px_shift: Union[int, float]
+    :param py_shift: int or float, shift for spectral index to reconcile sky
+        wavelength with fit parameters, will be added to spec to get py in theta
+    :type py_shift: Union[int, float]
+    :param float alpha_min_index: float, index of the minimum grating angle
+        computed from the reference alpha
+    :param float c0: float, fitted coefficient
+    :param float c1: float, fitted coefficient
+    :param float c2: float, fitted coefficient
+    :param float c3: float, fitted coefficient
+    :param float c4: float, fitted coefficient
+    :param float c5: float, fitted coefficient
+    :param float c6: float, fitted coefficient
+    :param float c7: float, fitted coefficient
+    :param float c8: float, fitted coefficient
+    :return: float or array of the wavelength(es) for the input spat and spec
+    :rtype: Union[float, numpy.ndarray]
+    """
+
+    alpha_s = alpha_min_index - grat_idx / (rd * rg)
+    sin_alpha_s = np.sin(alpha_s * np.pi / 180)
+    px, py = spat + px_shift, spec + py_shift
+    theta = px + c6 * py ** 2 + c7 * py + c8
+
+    a = c5
+    b = c0 * theta + c4
+    c = c1 * theta ** 2 + c2 * theta + c3
+    wl = 5 / order * (a * sin_alpha_s ** 2 + b * sin_alpha_s + c)
+
+    return wl
+
+
+def wl_to_spec(wl, spat, grat_idx, order=5, rd=200, rg=0.711111111111111,
+               px_shift=0, py_shift=0, alpha_min_index=73.9062453,
+               c0=0.9800910119, c1=-0.0017009051035, c2=-0.87654448327,
+               c3=36.248043521, c4=459.42373214, c5=-80.04474108,
+               c6=-0.0017003774252, c7=-1.5498032937, c8=102.04705483):
+    """
+    inverse function for grat_to_wl(), compute spectral index for input spatial
+    position, wavelength at the given grat_idx
+
+    :param wl: float or int array, wavelength used to compute spectral index; if
+        input wl is an array, it must be compatible with spat so that spat+wl
+        does not raise an error
+    :type wl: Union[float, numpy.ndarray]
+    :param spat: int or int array, spatial position considered; if input spat is
+        an array, it must be compatible with spec so that spat+wl does not
+        raise an error, and the output array shape will be the same as the
+        shape of spat+spec array
+    :type spat: Union[int, numpy.ndarray]
+    :param int grat_idx: int, grating index
+    :param int order: int, grating order to use
+    :param rd: int or float, stepper motor, degrees per step
+    :type rd: Union[int, float]
+    :param float rg: float, grating degrees of movement per rotation of drive
+        shaft
+    :param px_shift: int or float, shift for spatial position to reconcile sky
+        wavelength with fit parameters, will be added to spat to get px in theta
+    :type px_shift: Union[int, float]
+    :param py_shift: int or float, shift for spectral index to reconcile sky
+        wavelength with fit parameters, will be added to spec to get py in theta
+    :type py_shift: Union[int, float]
+    :param float alpha_min_index: float, index of the minimum grating angle
+        computed from the reference alpha
+    :param float c0: float, fitted coefficient
+    :param float c1: float, fitted coefficient
+    :param float c2: float, fitted coefficient
+    :param float c3: float, fitted coefficient
+    :param float c4: float, fitted coefficient
+    :param float c5: float, fitted coefficient
+    :param float c6: float, fitted coefficient
+    :param float c7: float, fitted coefficient
+    :param float c8: float, fitted coefficient
+    :return: float or array of the spectral index(es) for the input spat and wl
+    :rtype: Union[float, numpy.ndarray]
+    """
+
+    alpha_s = alpha_min_index - grat_idx / (rd * rg)
+    sin_alpha_s = np.sin(alpha_s * np.pi / 180)
+    px = spat + px_shift
+
+    a = c1
+    b = c0 * sin_alpha_s + c2
+    c = c3 + c4 * sin_alpha_s + c5 * sin_alpha_s ** 2 - wl * order / 5
+    theta = (-b - np.sqrt(b ** 2 - 4 * a * c)) / 2 / a
+
+    a = c6
+    b = c7
+    c = c8 + px - theta
+    py = (-b - np.sqrt(b ** 2 - 4 * a * c)) / 2 / a
+    spec = py - py_shift
+
+    return spec
+
+
+def wl_to_grat_idx(wl, spat, spec, order=5, rd=200, rg=0.711111111111111,
+                   px_shift=0, py_shift=0, alpha_min_index=73.9062453,
+                   c0=0.9800910119, c1=-0.0017009051035, c2=-0.87654448327,
+                   c3=36.248043521, c4=459.42373214, c5=-80.04474108,
+                   c6=-0.0017003774252, c7=-1.5498032937, c8=102.04705483):
+    """
+    compute the required grating index to place the input wavelength at the
+    desired spatial position and spectral index
+
+    :param wl: float or int array, wavelength used to compute grating index; if
+        input wl is an array, it must be compatible with spat and spec so that
+        wl+spat+spec does not raise an error, and the output array shape
+        will be the same as the shape of wl+spat+spec array
+    :type wl: Union[float, numpy.ndarray]
+    :param spat: int or int array, spatial position used to compute grating
+        index; if input spat is an array, it must be compatible with wl and spec
+        so that wl+spat+spec does not raise an error
+    :type spat: Union[int, numpy.ndarray]
+    :param spec: int or int array, spectral index used to compute grating index;
+        if input spec is an array, it must be compatible with wl and spat so
+        that wl+spat+spec does not raise an error
+    :type spec: Union[int, numpy.ndarray]
+    :param int order: int, grating order to use
+    :param rd: int or float, stepper motor, degrees per step
+    :type rd: Union[int, float]
+    :param float rg: float, grating degrees of movement per rotation of drive
+        shaft
+    :param px_shift: int or float, shift for spatial position to reconcile sky
+        wavelength with fit parameters, will be added to spat to get px in theta
+    :type px_shift: Union[int, float]
+    :param py_shift: int or float, shift for spectral index to reconcile sky
+        wavelength with fit parameters, will be added to spec to get py in theta
+    :type py_shift: Union[int, float]
+    :param float alpha_min_index: float, index of the minimum grating angle
+        computed from the reference alpha
+    :param float c0: float, fitted coefficient
+    :param float c1: float, fitted coefficient
+    :param float c2: float, fitted coefficient
+    :param float c3: float, fitted coefficient
+    :param float c4: float, fitted coefficient
+    :param float c5: float, fitted coefficient
+    :param float c6: float, fitted coefficient
+    :param float c7: float, fitted coefficient
+    :param float c8: float, fitted coefficient
+    :return: float or array of the grating index(es) for the input spat, spec, wl
+    :rtype: Union[float, numpy.ndarray]
+    """
+
+    px, py = spat + px_shift, spec + py_shift
+    theta = px + c6 * py ** 2 + c7 * py + c8
+
+    a = c5
+    b = c0 * theta + c4
+    c = c1 * theta ** 2 + c2 * theta + c3 - wl * order / 5
+    alpha = np.arcsin((-b + np.sqrt(b ** 2 - 4 * a * c)) / 2 / a) * 180 / np.pi
+    grat_idx = -rd * rg * (alpha - alpha_min_index)
+
+    return grat_idx